feat: utils/constants.ts -> types/index.ts 로 타입 관련 코드 이동 (#42)
