+++ src/api/user/user.controller.ts
@@ -8,9 +8,10 @@ import {
 } from '@nestjs/swagger';
 
 import { UserService } from '@/api/user/user.service';
-import { ITokenUser, User } from '@/decorators/user.decorator';
+import { User } from '@/decorators/user.decorator';
 import { User as UserEntity } from '@/entity/user.entity';
 import { JwtAuthGuard } from '@/guard/jwt-auth.guard';
+import { TTokenUser } from '@/types';
 
 @Controller('user')
 export class UserController {
@@ -27,7 +28,7 @@ export class UserController {
   @ApiOperation({ summary: '유저 정보 가져오기' })
   @UseGuards(JwtAuthGuard)
   @Get()
-  async getUser(@User() user: ITokenUser) {
+  async getUser(@User() user: TTokenUser) {
     return this.userService.getUser(user.id);
   }
 }
